Add files via upload
**Design, Analysis, and Algorithms in Java**

This guide provides comprehensive details of lab programs, including explanations and outputs, which are useful for semester finals and lab exams. It's beneficial for all B.Tech students or anyone studying in this domain. By writing the Aim, Description, Program, and Expected Output clearly, you will perform well in lab exams and earn good marks in semester exams when asked to write a program.